refactor: extracts mouse activity pattern
